requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-openai>=0.1.0
pydantic>=1.10.8
reportlab>=4.0.0
wheel>=0.45.1
packaging>=24.2
toml>=0.10.0

# LLM providers
openai>=1.6.0
anthropic>=0.6.0  # Optional: for Claude support

# PDF generation
reportlab>=4.0.0

# Visualization (optional)
matplotlib>=3.7.0  # Optional: for generating visualizations
numpy>=1.24.0     # Optional: for data manipulation

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Web Research and Browser Automation
firecrawl-py>=0.1.0  # For web crawling and research
selenium>=4.10.0     # For browser automation
webdriver-manager>=3.8.5  # For managing webdriver installations
